Fix grammar that allowed forbidden component updates that the rest cannot handle
Accidentally the former syntax category "subscription" that was used in the
LHS of assignments, and which should only allow identifiers, was merged with
the more liberal variant that is allowed as general expression, before the
extension of subscription and slicing were done. The resulting too loose
grammar easily allowed crashes so the distinction had to be restored. This is
somewhat painful since the limited look-ahead forces to let the categories
IDENT and 'comprim' live alongside in numerous near-duplicate rules in order
to avoid parsing conflicts. Hopefully similar duplication of large swaths of
rules will not be (for other reasons) be needed again in the future.