Add realex function |swiss_matrix_knife| and matrix slice syntax that uses it
Like the by-row matrix syntax [1,2,3|4,5,6] which implicitly calls transpose,
the matrix slice syntax A[ i:k, j:l ] with optional bound reversals in the 4
positions (and the option to omit a bound entirely) is transformed by the
parser into a function call; in this case to a new function
|swiss_matrix_knife|, which has even 4 more option bits (whence the name) that
can select reversal of rows, columns, transposition, and negation of entries.

Also included discussion of the recently added features in realex.help.