Create option -Dincompletecpp11 to control old-style. Make |expr| mostly POD.
The compiler option is used to specify old-style work arounds. However this
change makes the |expr| type almost entirely a POD type even when this option
is absent, since using smart pointers in a |union| is not really worth the
hassle. Only one field that is |std::string| remains as non POD component.