Make the realex program compile and execute correctly with g++-4.4 and g++-4.6
It also still compiles and executes correctly with gc++-4.8.2. This was a
rather delicate task, having to navigate between the obligation to declare the
copy constructor of type_expr deleted (without which it sneaks in at return
statements (that one one would have thought elided) leading to memory
corruption, and on the other hand the necessity for |std:map| instances to
have a value type with a copy constructor, even if only for the trivial case
of a value (component) that has just been default-constructed.