Allow component update assignments like 'v[i]+:=1'; print type () as void
An assignement 'a[E] OP:= V' is translated to 'let $=E in a[$]:=a[$] OP V'
where '$' is a secret identifier. This necessitated for parsing reasons
contracting in lexer.w any operator with a following ':=' to a single token.

In passing 'quiet' and 'verbose' lose the status of keywords, and the commands
to use them become 'set queit' and 'set verbose'.