Add realex syntax for direct recursive functions, and handle |not| differently
The syntax for a recursive function is 'rec_fun f (T1 x)T2: body' where 'f' is
the recursive funtion name, 'T1' and 'T2' are argument and result types (both
obligatory) and 'body' is the function body that may call 'f'. This syntax
contains no '=', though often 'set f =' (with again 'f') will precede it.

The handling of boolean negation now remains represented as such in the syntax
tree (rather than being replaced by a conditional) and after the type-check
becomes a call of a new built-in function that is printed as '{!@bool}'.