Separate method |involutionMatrix| into |matrix| and |compute_matrix|
In the |ComplexReductiveGroup| class, the |involutionMatrix| always computed a
reult on the fly from distinguished involution and twisted involution. This is
ncessary during construction, but afterwards is better handled by looking up
the matrix in the involution table. Now two different methods are provided
(with |matrix| returning a reference) and proper distinction is made at uses.