Simplify method implementing cross@(vec,Param): avoid modifying |gamma| at all
An initial transformation to the situation where the infinitesimal character
representative is integrally dominat was done, and an inverse transformation
afterwards. But for cross action this is not necesary: the code for the
transformation will work directly. Although the code for computing |lambda|
uses the halfsum of the positive real roots, its effect is independent on
which notion of "positive" is used to determine it; whence the equivalence.