Make plan for renaming of the executable names
The 'atlas' executable will be renamed to 'Fokko'
The 'realex' executable will be renamed to 'atlas'
The Atlas-independent programming language currently incorporated into the
'realex' program will be named 'axis'; no seperate program of this name will
be developed for now, to this is name is mainly for internal usage only.
The '.rx' default name suffix for scripts is to replaced by the '.at' suffix.