Quick bugfix, remove illegal comparisions |file==std::cout| (gcc 5.3 found it)
In realmode.cpp, the code was trying to test whether an OutputFile referred to
|std::cout| by testing the refernece held against |std::cout|. This can never
have done what was intended, which should have compared pointers; it is
unclear just why this passed compilation for many years. The quick fix was to
add and call a specific mehtof to |OutputFile| for this. For the concrete
situation, having a constructor that forces disk file would be a better fix.