Define |RealReductiveGroup::base_grading|, and test that |KGB::KGB| uses this
This method stadardises the coweight used in a square class (if only, at
present, to define the base grading) through |realredgp::square_class_choice|,
and to make the test succeed, |complexredgp::square_class_grading| which is
currently used, had to be made to normalise as well (this makes it more
complicated, but the function is planned for removal). As a consequence, the
base grading and the choice of the element KGB(G,0) change (within its KGB).