Modify type of |cocharacter| stored in |real_form_value| to |RatCoweight|
This used to be |TorusElement|, in other words reduced module 2\Z^n, but that
representation is unsuited to control hagin a $\xi^t$-stable lift, and it does
not match the type stored in in |RealReductiveGroup| that will replace it.

Code does not pass test now, but this seems to be due to |strong_real_form_of|
doing complicated manoeuvres that are to be removed anyway.