Resuscitate old function |real_form_of|, but now also return |cocharacter|.
This is to replace |strong_real_form_of|. The idea is that the nasty part of
the latter function tries to do something it shouldn't, namely to distinguish
different strong real forms associated to a same weak real form by twiddling
the cocharacter. But the cocharacter should just depend on the square of the
strong involutions, so the distinction here should be in the torus bits for
the intial KGB element |x0| for the real form instead.

Getting the |real_form_of| function right required adding a |stable_log|
function to find a $\xi$-stable rational coweight representing a torus element
(in practice the shited square for the square class). It is very similar to
|square_class_choice|, and it seems that former could serve for the latter.