Implement "normalise before using table" for extended parameters in length 1
Concretely the procedure |type| that computes the type of an extended
parameter adjacency for a given extended generator, and computes the
neighbouring extendes parameters as a side effect, now does this by first
adapting the original extended parameter to that adjacency (so that the table
for cross/Cayley is supposedly reliable), and exports the sign involved in
making the change. Adaptation is currently only done for the length 1 case.