Make |overloaded_builtin_call| have poitner to |builtin_value|, sharing string
The string that is stored in an |overloaded_builtin_call|, representing the
function name and its argument types and used for trace-back if an error
should be thrown by the built-in, used to be duplicated whenever a function
call was compiled. This is now avoided by storing a shared pointer to the
|builtin_value| that represents the built-in, most often held in the global
overload table (but being shared pointer it does not depend on that table).