Implement multi-way assignment, as in 'set (a,b):c := v'
This is complicated by syntactic considerations (unwarned, the parser has
problems recognising the left hand pattern, whence the 'set'), and type
analysis and evaluation are also more difficult due to a possible mix of local
and global destination variables. All in all this complements rather than
supersedes the possiblilty of simple assignment statements.