Set up administration of |in_loop| attribute inside |layer| structure
To get this working we needed to introduce empty |layer| instances for while
loops and four counted loops without named loop index, and to match that also
empty |frame| instances at runtime. To prevent the extra cost that this incurs
we should rework things so that a |layer| without identifier will not be
counted when measuring the |depts| of local idenifiers, and that no
corresponding |frame| will be generated.