Eliminate empty frames, thus resolving inefficiency of previous commit
To avoid that an empty static layer, created for the sole purpose of switching
the "inside a loop" status needed to check the validity of |break|, slows down
local identifier location at runtime, it is now stipulated that such layers
are not counted in the depth of a local identifier, and that there will be no
corresponding stack frame at runtime. While the explicit empty frames
introduced in the previous commit for while loops were easy to remove, we also
had to take care to eliminate all possible "accidental" occurrences of empty
frames. This was mostly done at compile time, transforming let expressions and
component-for-loops guilty of empty layers into respectively sequence
expressions and counted for loops. There remain some dynamic tests for calls
of user defined functions, inevitable in the general case, while for the case
of overloaded calls it could in principle be resolved at compile time instead.