Re-implement value-gathering while-loops using a |containers::simple_list|
This used to be done using a |std::vector|, with consequent re-allocations and
reversals being necessary. This change probably gives a negative speed-up, but
it is in any case only small, might become positive if faster allocators can
be found for lists, and allows us to gives our lists some serious exercise.