Clean up some scripts that used suffix/join operator # with LHS of type [*]
While axis resolved this potentially ambiguous case (actually ambiguous if the
RHS has some row type) in favour of the suffix interpretation, this ambiguity
is potentially harmful and moreover inconsistent with the rule that [*] only
matches in overloading if explicitly given as type. In order to prepare for
removing this laxness from the language, the occurrences in the scripts library
are improved by casting empty lists (at initialisation) to the proper types.

We also cleaned up one case where an operand of |#| was assumed to undergo
coercion, as the rule will be that matching generic types excludes coercion.