Limit insertion of runtime "voided" coercions to certain exotic cases.
The old strategy was to insert a |voiding| node in the converted program every
time an expression of non void type was contextually converted to void type.
This ensured that no value was ever generated from such subexpressions, but in
most cases, including all cases that normally arise such as an assignment
followed by a semicolon, the |voiding| node is actually redundant: the
position of the subexpression in the converted program already ensures that it
will be evaluated without producing a value. The change made consists of
shifting the responsability of inserting a |voiding| from |conform_types| to
some places in |convert_expr|, inserting one only in those cases where the
latter detects a void type for an expression without the syntactic context
implying this, such as the RHS of an assignment where the variable turns out
to have void type. Such cases are invariably extremely rare in practice.