Fix bug in construction folded Dynkin diagram; make 'braid' use repr-mode
Having the 'braid' command (Fokko) use repr-rmod means that it can be used to
test the new code that flips edges based on extended paraeter computations.

The but in dynkin.cpp was silly: between orbits of different sizes that were
detected to be neighbours, the code forgot to set the bits indicating this.