Fix a misunderstanding in the 1r case of parity condition in A2-like case
For a delta-fixed non-integral real root that cannot be conjugated to a
delta-fixed real simple root, there was confusion about how to test the parity
condition, and how to update |rho_r_shift|. The conjugation under $W^\delta$
leads to a delta-fixed sum of two delta-switched simple roots (A2 situation),
and for the parity condition the conjugating word should conceptually be
prefixed by a conjugation of that sum to one of its simple summands. This
changes the parity condition if and only if the summands are real (rather than
complex). Correspondingly, a correction to |rho_r_shift| by one of the
summands should be made only if the summands are real. This condition is
equivalent to the type 2 condition that alpha is not in the image of $1-\theta$
(when the summands are complex, $1-\theta$ maps each of them to $\alpha$).