Improve handling of 2Ci/2Cr cases for non-integral blocks; not yet perfect.
In non-integral blocks the two complex roots involved in a 2Ci or 2Cr
situation need not be simple, and when they are not, the computed extended
parameter sometimes failed to lie over the correct ordinary parameter, in
other words the computation was incorrect already for |lambda_rho| or |l|.
(For instance 'D5 ad c', x=244 (max), lambda=rho, nu=rho/4 exemplifies this.)

The problem is caused (for $\lambda$) by the fact that cross action by a
non-simple complex root does not map $\rho_r$ at the original involution to
$\rho_r$ at the destination involution (breaking up the cross action into
simple ones, there may occur cross actions by real roots, which introduce a
shift for $\lambda$). This was corrected by adding the difference, the sum of
positive real roots at arrival that are images of negative roots and which is
conveniently computed by |repr::Cayley_shift|, to the |lambda_rho| component
of the extended parameter, and to do a dual adaptation for the |l| component
(this required introducing a dual version |ell_shift| of |Cayley_shift|).

Since these corrections have no reason to be $\delta$-fixed, an adaptation of
the |tau| and |t| fields are also necessary. Providing this correction by
adding half the $(\delta-1)$-image of the correction to $\lambda$, as
suggested in Lemma 10.5, does not work since the division by 2 is not exact.
Therefore instead we now invoke |matreduc::find_solution| to provide valid
corrections to |tau| and |t|. However, this relinquishes control of the
sign of the resulting extended parameter, so braid relation failures subsist.