Simplify quotient of |z| value function to ignore second term in |z| formula
After long consideration of what to do for a Cayley transform by an integrally
simple but not simple root (the previous code gave a braid failure for 'D7
2/4 c' then trivial representation "scaled" by 1/3, but the problem already
occurs for SL(3,R)), the verdict was that the quotien of |z| values should not
be evaluated at the non-simple situation, but after conjugation to a simple
situation. But in that situation the second term contributes nothing, so the
code was changed to behave as if there were no second term in the first place.

For the record, there is still a braid failure for 'D6 1/2,1/2 c', then
trivial*/2. This failure appears to be independent of the problem fixed here.