Reestablish |z_quot| contribution from second term in |z| for specific cases
The argument that the second (exponent of -1) term in |z| never affects the
sign was flawed; indeed the correct handling of 2i12 cases hinges on it. What
is needed is a factor minus one to the power |t*d_lambda| where |t| is the
adjusted-t-downstairs value of |t|, and |d_lambda| is the change to
|lambda_rho| (if any) that does not come from |Cayley_shift|. In many cases
this pairing is automaically zero, including all length 1 and length 3 cases,
but several cases in length 2 do need this to be computed and adjusted.