Fix two errors inherited from 'tinker': extended length and use of |Cayleys|
Once again a definition of length in extended blocks that differs from the
length in the parent block threatened to infect the code base: having been
corrected in |extended_block| (now phased out) it had bee reintroduced 9 Sep
2015 (commit 3e3ac8) in the |ext_block| class; this set straight (equal to
parent) again now. The other error was adaptation to a specification
difference between |extended_block| and |ext_block|: the method |Cayleys| of
the latter should not be called unless the transform is actually two-valued.

Some minor code embellishments were also copied from a to-be-abandoned branch