Move some methods only used for only one block class into that (derived) class
There was confusion about ensuring that all necessary block fields are
initialised. The method |element| was virtual in |Block_base|, but in fact
only ued from |Block|. The method was made a non-virtual one of |Block|, and
the |first_z_of_x| member it uses was also moved there. The |sort_by_x| method
is only used from the |param_block| constructor, so it was moved from
|Block_base| to that class. But in fact sorting a |param_block| by |x| values
is not essential without the |element| method, so the method may change soon.
In fact we prepare for possibly sorting |param_block|s by length instead.