Change generation of |param_block| so that it sorts by length first (then |x|)
It is a basic assumption of |Block_base| that the length of elements is weakly
increasing along the block. This was destroyed (for non-integral parameters)
in the |param_block| constructor by sorting globally by |x|. The generation is
actually by reverse length; now the reversal of the length is combined with a
local sorting by |x| values for ranges with constant length.