Fix two bugs in ext_kl.cpp: wrong block list, and wrong handling defect cases
The parameter list that |extended_KL_block| generated was of the correct size
but taken from the full block rather than only from the delta-fixed elements;
the passing to singular descents reduction overlooked the defect cases by
classfying all descents as like-noncompact, complex, type 1 or double valued.
For the latter problem, the right thing is simply to distinguish 0,1,2-valued.