Update |Rep_table::add_block| for extended blocks to use new functionality
The most vital change was to insert the edge signs on the extended block at
construction, which was an implict consequence of a previous commit. The code
of the method istelf is now adapted to recent improvements of defining what
"nonzero final" is in an extended block, and of condensing tables accordingly.