Implement values for union types, incidentally reorganise header inclusions
The implementation is quite simple, but makes axis.h need the |id_type|
typedef, which it did not have access to. The situation was muddy, so the
typedef is now made in buffer.h which everybody (including axis.h) can include
without much harm. The dependency of global.h on parsetree.h was then relaxed
to just buffer.h, which brought to light some places depending implicitly on
the former over-inclusion of header files; these have been set straight.