Fix bug in built-in |twisted_full deform|: don't lookup without any deformation
The code was depending on the current parameter being added to the tables when
the first defrmation is being computed, but this does not happen in the case
where there are 0 reducuibility points, as then nothing gets defoemed at all.
This is solved by not looking up values when there are 0 reducuibility points.