Let |param_block| methods |lambda_rho| and |lambda| make standard choices
The standard choice here is defined as the one for the (unique) |StandardRepr|
value corresponding directly to the block element. The previous |lambda_rho|,
which is now a private |internal_lambda_rho|, also made a standard choice, but
a different one. This is a serious concern for the default choice of extended
parameters, which previously used different values for |lambda_rho| depending
on whether using a block (as in |extended_block|) or just a Param value (as in
|finalize_extended|. This patch should eliminate such differences. The way it
is implemented now is a cluch; much better is to redo |param_block| internals.