Add |y_bits| field for |param_block| elements, as in atlas |Param| values
This is the beginning of a reorganisation of the |param_block| structure to
eventually do away with the |TorusElement| values stored to represent the |y|
component of block elements, and to relegate them to block construction only.

Currently the |y_bits| are generated and compared to see if they produce
identical |lambda_rho| values to the current ones, through the method
|new_lambda_rho|. The next step is to replace the |TorusElement| based
accessor methods by ones that are based on the |y_bits| values.