Remove |old_lambda_rho| and its acolites, and make |param_block::xsize| correct
The point of the latter is that since |x| is no longer guaranteed to be
increasing in a |param_block|, one cannot simply take the final |x| value as
an upper limit for the |x| values that occur, as |xsize| assumed it could.

Also replace |xsize| and |ysize| virtual |Block_base| methods, which only
serve for polymorphic printing, by methods |max_x| and |max_y| that are more
appropriate for this use, and eradicate all other, unnecessary, uses.