Fix |KGB::twisted|, remove version in ext_block, remove unused arguments
The |KGB::twisted| method forgot to take into account the action of the
external |delta| on the real form base (grading vector). This was done
correctly in the version of the same functionality in ext_block, whose code
was borrowed from but which was then removed and |KGB::twisted| used instead.
The change also brought to light that several function had arguments that were
or became unused (typically a separate |weyl::Twist|), so these were removed.