Replace hash tables in param_block |TorusElement| based to |StandardRepr| based
Rooting out the dependency on |TorusElement| storage in |param_block| required
many changes. The hash tables that were permanently in |param_block| are now
local variables to the constructors, and references are needed to be passed to
auxiliary methods (used during construction only) that no longer have access
to the automatically. The result does not look nice, but indicates more
clearly how the construction functions inter-depend. One not entirely clerical
change is that the entry element lookup (full block) now uses |StandardRepr|.
Another is that |reverse_length_and_sort| now leaves reconstructing the block
hash table to the caller, so that it does not need an argumnet just for that.