Have |scale_extended| do complex singular descents too, "half descent" for 2Ci
Not doing complex singular descents was causing problems in
|twisted_full_deformation|, since the initial |make_dominant| in |add_block|
was causing the a smaller block to be added than necessary to find the passed
parameter in the tables afterwards.

If a complex singular simple root is of type 2Ci, then trying to do an
extended complex descent through its orbit would bounce back to the original
extended parameter, and hang in a non-advancing loop. Instead we do like
|star| does for a 2Cr, mostly reflecting by one simple root in the orbit, but
adapting to |tau| for the non-$\delta$-fixed change made to |lambda_rho|.

The code now asserts that a similar situation for 3Cr cannot occur, but upon
scaling by a factor 0 that does seem possble, at least with external $\delta$