Add function |Cayley_shift_flip| in ext_block.cpp, used when |Cayley_shift| is
Every time a conjugation to (almost) simple takes place, we now compute the
sign of the permutation $\delta$ defines on the set of positive roots made
negative by the conjugation that in addition change their real status from
upstairs to downstairs, and the sign is recorded in extended parameter flip.

This makes a slight change to GL(4,R), but does not make its trivial unitary.