Start simplification ext_block: |star| records flip in |param| value itself
Thus the |star| output no longer produces a sign-parameter pairs: the flipped
status is recorded in the extended parameter output itself. This uses in most
cases a new function |z_align| instead of |z_quot|; instead of computing a
flip value, in records what |z_quot| would have recorded in the flipped status
of its second argument parameter. This being a setter of that status, the old
value of same is not used. However some information was sometimes put in that
status before calling |z_quot|; this now needs to be passed separately to
|z_align| that has an additional |bool| argument for the purpose.

However in |complex_cross| and in the 2Ci and 2Cr cases, |z_align| is not
called, and in this case the flipped status of the input |E| to |star| must
therefore be explicitly propagated to the output parameters |F| in the
links that |star| produces, these are the only cases where |F| is passed a
flipped status on construction.

Correct propagation of flipped status in |complex_cross|