Rewrite |complex_cross| to act one extended parameter by shifted reflections
This used to be action on individual components that were shifted before and
after to give the effect of affine reflections. The code now uses the new
affine reflection primitives in the |RootDatum| class, for which purpose the
|ext_block::param| structure had to make its fields public (the affine
relection operations are only defined to work in-place). This should be made
more proper again in a future change, for instance using a friend declaration.