Fix errors in |z_align| and in affine reflection implementations (rootdata.h)
The |z_align| function should not ignore the flipped status of first extended
parameter, but proagate it into the second. This bug was introduced by the
change that eliminated the explicit signs in the output from |star|; for
formal equivalence with the old situation, both flipped statuses and the
quotient of |z| values should all be incorporated into the sign that is stored
as the new flipped status of the second parameter. The affine reflections had
a silly bug in the implementation |shift[i]| sould be |shift[ww[i]]| (twice).