Replace |Cayley_shift_flip| by more general |shift_flip| and some auxiliaries.
We now always take the XOR of real positive roots upstairs and downstairs to
filter the roots mapped from positive to negative when conjugating the
integrally-simple root alpha to a simple root. In Cayley links, this should be
the same as just filtering by positive roots upstairs, but the new
functionality has potential new applications, that are not yet exploited.