Add possibility to act on |gamma| of |ext_block::context|, and use it
This is for use in |scaled_extended_dominant|. It makes it possible to
actually have an extended parameter to work on, rather than simulating it by
decomposing it in components. A next step is to replace the 2Cr simulating
code by an actual call to |star|, avoiding code duplication and divergence.