Change |scale_extended| to call rather than emulate |star| for 2Cr descents
The function |scale_extended| used to emulate the 2Cr section of |star| when
it needed to do a singular complex descent for a simple root in a length 2
orbit for delta. This was motivated among other things by the fact that there
was no true extended parameter at hand, the main work of |scale_extended|
being modification of |gamma| which is not stored in the (extended) |param|
value but in the |context|. However a lot of details of |star| were not
faithfully copied, and in spite of previous corrections the code replaced in
this commit still did not include the final |sign| computation of |star| for
the 2Cr case (nor the October surprise correction at the very end of |star|,
but this actually does nothing for 2Cr links).

To make such clerical errors impossible, a call is now made to |star| to do
the actual work. This does require looking up the integral generator orbit
that contains the simple root in question (the simple root being sigular, it
must be integrally simple) since |star| wants such an orbit as argument.