Revert |scale_extended| to not use extended parameter initially, but do at end
Doing the part making |gamma| dominant using a |param| and separately updating
its |context| to record the chinging |gamma| value proved to be problematic,
since the other field of the |context| (notably the integral datum and the
subsystem) did not follow. Rather than implement the changes that would be
necesssary to make this work corectly, we reverted to the idea of first
working on isolated components of an extended parameter, and a |gamma| value,
then once |gamma| is dominant build a |context| and a true |param| value.
This requires separating out the (now) final phase of descending through
(extended links corresponding to) singular complex descents. The latter
(still) uses |star| to ensure that signs on the links traversed are exactly as
they are when constructing an extended block. This is somewhat awkward, since
contrary to the initial part it must be done in terms of (singular) generators
for the integrality datum rather than of the full datum.

Currently there appears to still be an error, in that SL(4,C).trivial fails to
report as unitary (as does SL(5,C).trivial).