Modify ext_block to approach the work David has done; making dominant l=2 flip
Main change this commit: do flip in the first "make gamma dominant again"
phase of |scaled_extended_dominant|, whenever a type 2 orbit reflection is
done. Other non-serious stuff: moving unconditional flips (for "wedge product")
in the 2i/2r cases, upward in the code to before splitting the cases out, and
adding a |check_quadratic| function to check quadratic relation in a block
(this function is not yet invoked in any way).