Incomplete attempt to fix 3Ci/3r link signs.
Goal is to find common values of lambda,tau,l,t that work for both
sides of the link. The formulas in the twisted paper don't do this,
but it should be possible. Once it's done, I think that the sign of
the link will be something extremely simple (perhaps always -1??)

For the moment the attempt to construct these value fails for GL(5,R)
at rho/2 for the 3r link from block element 7 to block element 0:
precisely, the new lambda constructed fails to agree with gamma-rho on
the +1 eigenspace of theta. I don't understand why.