Change the formula for flip due to a integrally-simple root being made simple
It still depends on the set of positive roots being made negative under the
sequence of reflections used to make the integrally-simple root simple, but
instead of counting two-element $\delta$-orbits on the set of roots that
change reality across the link to those that change very-complexity, the
latter meaning the circumstance that $\theta$ sends the root outside its orbit
under the 4 element group $\pm1,\pm\delta$. Orbits under $\delta$ of two roots
whose sum is again a root are excluded from the count.

Since the |rho_r_shift| remains equal to the sum of the |pos_to_neg| roots
that change reality, this change required some internal reorganisation.

There is now one central function |shift_flip| doing the computation, which is
called from |complex_cross| as well as from various place in |star|.