Thoroughly reorganise |scale_extended|: separate stages, 2nd one uses |star|
The function |scale_extended| used to emulate the 2Cr section of |star| when
it needed to do a singular complex descent for a simple root in a length 2
orbit for delta. This was motivated among other things by the fact that there
was no true extended parameter at hand, the main work of |scale_extended|
being modification of |gamma| which is not stored in the (extended) |param|
value but in the |context|. However a lot of details of |star| were not
faithfully copied, and in spite of previous corrections the code replaced in
this commit still did not include the final |sign| computation of |star| for
the 2Cr case (nor the October surprise correction at the very end of |star|,
but this actually does nothing for 2Cr links).

To make such clerical errors impossible, a call is now made to |star| to do
the actual work. But |star| needs an actual extended parameter, while using
that in the part making |gamma| dominant would mean changing its |context| to
record the changing |gamma| value; this would require updating other fields of
|context|, which would need appropriate methods in the classes involved like
|SubSystem|, a major overhaul. So instead the making |gamma| dominant part was
separated from the singular complex descent part. The former works on
isolated components of an extended parameter and a |gamma| value, and once it
is dominant a |context| and a true |param| value can be built. After this the
latter part can be done, using a call to |star|.

The second part remains somewhat awkward, since contrary to the initial part,
|star| operates in terms of (singular) generators for the integrality datum
rather than of the full datum. So we need to look up the integral generator
orbit that contains the simple root in question. Since that root is singular,
it must be integrally simple, and the search must succeed.

Also a flip for 2C cases in first phase, and a conditional flip to undo
possible October surprise flip from |star|, in the second.