Do flip for both Cayleys in cases 1i2f and 1r2 when alpha cannot be made simple
In the 1r2 case, more precisely when |shift_correct| holds (a subcase of
failure to reflect alpha to a simple root, where in addition alpha is the sum
of two real roots). The |shift_correct| condition implies type 2 (dixit
David), which explains the absence of any conditional flipping in the length 1
type 1 cases.