Rewrite |complex_cross| to act on extended parameter by shifted reflections
This used to be action on individual components that were shifted before and
after to give the effect of affine reflections. The code now uses the new
affine reflection primitives in the |RootDatum| class, taking advantage of the
fact that the |ext_block::param| structure has made its fields public (the
affine relection operations are only defined to work in-place).