Fix the fact that user interrupts are not cleared by e.g. 'set' commands
The interrupt flag used do be cleared before evaluating an expression, but
this misses any command that is of a different type, notably "set ... = ..."

This is fixed by clearing the |interrupt_flag| as soon as it is found nonzero,
which always happens in the axis.w module. Some relevant code was moved there.