Separate "dominant" from "normal" conditions for parameters
Internally, there are now separate methods |make_dominant| and |normalise| in
the |Rep_context| class; the latter calls the former before (for now)
decending through singular complex descents (a next task will be to insert a
unifying operation to ensure that |normalise| really gives a unique result for
all paramters in an equivalence class, which is currently not the case).

There were many internal calls of |make_dominant|; some but not all of these
have been replaced by |normalise|. Those that we decided need not be replaced
may get a slight defiierence in behaviour, but we deem that any code that
should exhibit the difference (which we hope is not present in the project)
was depending on non guaranteed implementation details anyway. In the same
vein, though not strictly using the same primitives, the built-in
Cayley@(vec,Param) does not do a full |normalise| for the integral system
internally before its actual action; since it transforms back to the original
situation anyway, this should actually not make any difference at all.

Finally the 'atlas' program now has separate |dominant| and |normal|
functions. The scripts were adapted to this (a single change in basic.at)