Redefine and distinguish equality (=) and equivalence of parameters
We now have two primitives, what used to be =@(Param,Param) is now called
equivalent@(Param,Param), and =@(Param,Param) is strict equality (this means
=@(Param,Param) has silently changed meaning, but it was not used anywhere in
the scripts). The implementation of equality was slightly improved, and the
occasion was take to make the interface to built-in equality operators somewhat
more regular (e.g., always defining an accompanying inequality operator).