Call |Rep_context::normalize| when making parameters from |StandardRepK| values.
The way |StandardRepK| is defined and the conversion is done, there is no
reason to assume that the parameters produced here, which end up unchanged in
the output |ParamPol| of |branch|, would be |normal| without such a call. This
change therefore should avoid the non-dominant parameters coming from |branch|.

In passing also cleaned up the comments in standardrepk.{h,cpp} a bit.