Added genuine.at, functions for testing K-type formulas
A ParamPol P is said to be "genuine" if it represents the K-types of a
form on an actual (not just virtual) representation. This condition is
hard to see in terms of the coefficients of P. It means that every
parameter occuring is a standardrepk, and if you convert the formula
to a formula in terms of actual K-types (finite dimensional
representations of K), then the multiplicity of each K-type is a+bs
with a,b\ge 0. Since there are infinitely many K-types this can be
checked only up to some height bound (there is a bound which suffices
to prove the formula is genuine).

The main commands are:
is_genuine:(ParamPol,int)->bool
test_genuine:(ParamPol,int)->(ParamPol,int)
is_hermitian_form_irreducible_genuine:Param->bool
test_hermitian_form_irreducible_genuine:Param->(ParamPol,bool)